// wwwroot/app/react/types/configurator.ts
export interface StageTask {
  taskId: string;
  required: boolean;
  config: TaskConfig;
  context: StepContext;
}

export type StageName = 'Enroute to Pickup' | 'Pickup' | 'Enroute to Delivery' | 'Delivery';

export type StagesMap = Partial<Record<StageName, StageTask[]>>;

export interface PresetWorkflow {
  name: string;
  stages: StagesMap;
}